# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.ErrorResponse:
    properties:
      details:
        additionalProperties: true
        type: object
      error:
        example: Bad Request
        type: string
      message:
        example: Invalid request body
        type: string
    type: object
  dto.SignUpRequest:
    properties:
      date_of_birth:
        example: "1990-01-01T00:00:00Z"
        type: string
      email:
        example: john.doe@example.com
        type: string
      first_name:
        example: John
        maxLength: 100
        minLength: 2
        type: string
      last_name:
        example: Doe
        maxLength: 100
        minLength: 2
        type: string
      password:
        example: P@ssw0rd
        maxLength: 72
        minLength: 8
        type: string
      phone_number:
        example: "+1234567890"
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  dto.SignUpResponse:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        example: john.doe@example.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      last_name:
        example: Doe
        type: string
    type: object
info:
  contact: {}
  description: This is a sample server for a Splitwise clone application.
  termsOfService: http://swagger.io/terms/
  title: Splitwise Clone API
  version: "1.0"
paths:
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Registers a new user
      parameters:
      - description: Signup Request
        in: body
        name: signupRequest
        required: true
        schema:
          $ref: '#/definitions/dto.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/dto.SignUpResponse'
        "400":
          description: Invalid request body or validation failed
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: User with this email already exists
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: User Signup
      tags:
      - Authentcation
swagger: "2.0"